+++ mycodo/inputs/atlas_orp.py
@@ -1,7 +1,10 @@
 # coding=utf-8
+import time
+
 from flask_babel import lazy_gettext
 
 from mycodo.inputs.base_input import AbstractInput
+from mycodo.utils.calibration import AtlasScientificCommand
 from mycodo.utils.system_pi import str_is_float
 
 
@@ -44,7 +47,6 @@ INPUT_INFORMATION = {
         'uart_location',
         'uart_baud_rate',
         'period',
-        'single_input_math',
         'custom_options',
         'pre_output',
         'log_level_debug'
@@ -64,13 +66,24 @@ INPUT_INFORMATION = {
 
     'custom_options': [
         {
+            'id': 'temperature_comp_meas',
+            'type': 'select_measurement',
+            'default_value': '',
+            'options_select': [
+                'Input',
+                'Math'
+            ],
+            'name': lazy_gettext('Temperature Compensation Measurement'),
+            'phrase': lazy_gettext('Select a measurement for temperature compensation')
+        },
+        {
             'id': 'max_age',
             'type': 'integer',
             'default_value': 120,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
-            'name': lazy_gettext('Calibration Max Age'),
-            'phrase': lazy_gettext('The Max Age (seconds) of the Input/Math to use for calibration')
+            'name': lazy_gettext('Temperature Compensation Max Age'),
+            'phrase': lazy_gettext('The maximum age (seconds) of the measurement to use for temperature compensation')
         }
     ]
 }
@@ -90,14 +103,19 @@ class InputModule(AbstractInput):
         self.i2c_bus = None
 
         # Initialize custom options
+        self.temperature_comp_meas_device_id = None
+        self.temperature_comp_meas_measurement_id = None
         self.max_age = None
         # Set custom options
         self.setup_custom_options(
             INPUT_INFORMATION['custom_options'], input_dev)
 
         if not testing:
+            self.input_dev = input_dev
             self.interface = input_dev.interface
-            self.calibrate_sensor_measure = input_dev.calibrate_sensor_measure
+
+            if self.temperature_comp_meas_measurement_id:
+                self.atlas_command = AtlasScientificCommand(self.input_dev)
 
             try:
                 self.initialize_sensor()
@@ -124,6 +142,32 @@ class InputModule(AbstractInput):
         """ Gets the sensor's ORP measurement via UART/I2C """
         orp = None
         self.return_dict = measurements_dict.copy()
+
+        # Compensate measurement based on a temperature measurement
+        if self.temperature_comp_meas_measurement_id and self.atlas_command:
+            self.logger.debug("pH sensor set to calibrate temperature")
+
+            last_measurement = self.get_last_measurement(
+                self.temperature_comp_meas_device_id,
+                self.temperature_comp_meas_measurement_id,
+                max_age=self.max_age)
+
+            if last_measurement:
+                self.logger.debug(
+                    "Latest temperature used to calibrate: {temp}".format(
+                        temp=last_measurement[1]))
+
+                ret_value, ret_msg = self.atlas_command.calibrate(
+                    'temperature', set_amount=last_measurement[1])
+                time.sleep(0.5)
+
+                self.logger.debug(
+                    "Calibration returned: {val}, {msg}".format(
+                        val=ret_value, msg=ret_msg))
+            else:
+                self.logger.error(
+                    "Calibration measurement not found within the past "
+                    "{} seconds".format(self.max_age))
 
         # Read sensor via UART
         if self.interface == 'FTDI':